설정 메인뷰 코드를 톱니바퀴 네비게이션 뷰로 이동
